Split huge test file into smaller files
The `hyperconverged_controller_test.go` is too large and freeze the IDE from time to time when editing it.
